requirements.txt
fastapi
uvicorn
python-dotenv
geopy
google-api-python-client
google-auth